+++ fugitive-darkness-api/src/main/java/com/zer0s2m/fugitivedarkness/api/handlers/ControllerApiGitProviderDelete.java
@@ -45,7 +45,12 @@ public class ControllerApiGitProviderDelete implements Handler<RoutingContext> {
                 .mapTo(ContainerInfoGitProviderDelete.class);
 
         gitProviderRepository
-                .deleteByTypeAndTarget(containerInfoGitProviderDelete.type(), containerInfoGitProviderDelete.target());
+                .deleteByTypeAndTarget(containerInfoGitProviderDelete.type(), containerInfoGitProviderDelete.target())
+                .onSuccess(ar -> gitProviderRepository.closeClient())
+                .onFailure((fail) -> {
+                    logger.error("Failure (DB): " + fail.fillInStackTrace());
+                    gitProviderRepository.closeClient();
+                });
 
         event
                 .response()
@@ -110,6 +115,8 @@ public class ControllerApiGitProviderDelete implements Handler<RoutingContext> {
                         } else {
                             event.next();
                         }
+
+                        gitProviderRepository.closeClient();
                     });
         }