+++ fugitive-darkness-provider/src/main/java/com/zer0s2m/fugitivedarkness/provider/SearchEngineGitException.java
@@ -1,0 +1,7 @@
+package com.zer0s2m.fugitivedarkness.provider;
+
+/**
+ * The main exception is for the search engine.
+ */
+public abstract class SearchEngineGitException extends Exception {
+}